Moved SEPADirectDebitTestHelper.java to demo package
