+++ Demo/src/main/java/com/braintreepayments/demo/SEPADirectDebitTestHelper.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.api;
+package com.braintreepayments.demo;
 
 import java.math.BigInteger;
 import java.util.ArrayList;
